bookstorelucky13/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
psycopg2-binary==2.9.4
python-dotenv==0.21.0
